wip: inter-stage variable trongle
this makes a _lot_ more sense now yeah. an @location is simply a
variable passed from vertex to fragment stage, and the @location(0) on
the fragment stage is the final color output. This on its own can lead
to some pretty fun stuff! If only I knew how to pass in triangles
dynamically...