wip: vertex buffers pt 2
very interesting, seems we can pass a lot of data with vertex buffers
that would otherwise just be indexed by instance. Both the vertex index
and the instance index can be removed this way! makes the shader itself
a bit conceptually simpler i think, setup is overall the same for it,
just have to know the difference between where/how you bind :)

There are definitely uses for storage buffers still, even in vs/fs
shaders, just not at this level of simplicity ig