attempt to scale & shift at the same time
Much harder than I had anticipated. The math involved is a lot tricker
than it appears at first glance, and even though I'm _pretty_ sure it
all checks out there's still some weird stuff happening when we need to
clip based on height (scale factor is off).

The "it bleeds into area it's not supposed to" is expected though, will
need to fix that somehow in the fragment shader perhaps. "if beyond the
bounds emit transparent" or something.