wip: flesh out more of the triangle render shader
still need to properly factor out the uniform calculation code into
something that can handle "fit" in addition to "fill". Also probably
want to have it so the starting scale can be specified more, right now
it is assuming a lot of things about being done directly from a texture.