fix camera calculations
turns out my previous code was assuming a 1x1 source always, and then
just using the passed-in rect for some sort of scaling calculation lol.
This worked for my texture shader (because it really is just always a
1x1 source for sampling the uvs properly), but this doesn't quite work
with my triangle shader. Modifying the camera to work directly on
vertices that are _actually_ in the source box of (0,0) to (w,h), and
then making the texture shader conform to this by reading the texture
size, works out just fine.

Had to modify the comment a little bit, but overall not too bad. Now I
have some abstract art in the upper right corner instead of a regular
GPU triangle! Good to know my triangle creation routines "just work" :)

Next up is figuring out the actual dimensions for the visualizer.