we have achieved wiggly lines!!
further things to consider:
+ i should probably bump the wiggles of the later bins too, by
  like 10x perhaps? & also scale their other impacts on parameters too
+ sqrt tuning on the levels? or like **0.9? idk exactly what the gamma
  should be, will again have to play around to find things that look
  right

next up is not actually that, but rather playback controls & progress
visualizer so I can test how this looks at different points in the
track, instead of just the beginning. kind of a requirement for later
experimentation lol, don't want to overfit on just this one.