connect to correct audio device on Linux
seems like KDE uses PulseAudio for its routing, at least as far as ALSA
(the only backend that cpal seems to support on Linux) is concerned.
Doing this lets the audio actually output to my headphones (finally!)
and can be switched dynamically on the fly with my system picker, neat.