wip: graphics for playback visualization
the graphics are not fully complete, but the necessary abstractions are
all there I think. factoring out the pipeline initialization was a bit
of a gamble and isn't really all that clean but what the heck it's a
quick & dirty app anyways.

Next up is moving the things it draws into the correct location, adding
the text properly, and piping the duration data down to it.